pyproject.toml
[project]
name = "tg-bot"
package-mode = false
version = "0.0.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "python-telegram-bot (>=22.0,<23.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "requests (>=2.32.3,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
